+++ game/physics/Physics_Player.cpp
@@ -1367,7 +1367,7 @@ bool idPhysics_Player::CheckJump( void ) {
 }
 
 //PN: Add jump
-void idPhysics_Player::addVel(int val, bool index) {
+void idPhysics_Player::addVel(float val, bool index) {
 	if (current.velocity[2] != 0 || index) {
 		idVec3 addVelocity = 2.0f * maxJumpHeight * -gravityVector*0.1*val;
 		current.velocity += addVelocity;
